+++ src/main/java/com/example/infraboxapi/orderItem/OrderItemDTO.java
@@ -22,7 +22,7 @@ public class OrderItemDTO {
     @Size(min = 1, max = 100, message = "Field 'name' must have a length between 1 and 100 characters")
     private String name;
     @PositiveOrZero(message = "Value must be a positive number or zero")
-    private Integer quantity;
+    private float quantity;
     @NotBlank(message = "Field 'item type' cannot be blank")
     @Size(min = 1, max = 100, message = "Field 'item type' must have a length between 1 and 100 characters")
     private String itemType;
